coin-ops/requirements.txt
werkzeug
json-rpc
jsonrpc-requests
cachetools